html/browsers/history/the-location-interface/location_href.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_href</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    if (('location' in window) && window.location != null
      && ('location' in document) && document.location != null){
      var href = location.href;

      test(function() {
        assert_equals(href, document.URL, "href attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("location interface is not supported.");
      });
    }
    </script>
  </body>
</html>
